fix(frontend): separate links and routes
