+++ packages/frontend/src/app/topics/[slug]/page.tsx
@@ -7,7 +7,7 @@ import { fetchTopic } from '@/fetchers/server/topic'
 // utils
 import { validateMeetingParams } from '@/utils/validate-meeting-params'
 // constants
-import { InternalRoutes } from '@/constants/navigation-link'
+import { InternalRoutes } from '@/constants/routes'
 
 export const dynamicParams = true
