+++ packages/frontend/src/components/topic/topic-feedback.tsx
@@ -6,7 +6,7 @@ import Link from 'next/link'
 import { colorGrayscale } from '@twreporter/core/lib/constants/color'
 import mq from '@twreporter/core/lib/utils/media-query'
 // constants
-import { InternalRoutes } from '@/constants/navigation-link'
+import { InternalRoutes } from '@/constants/routes'
 
 const Feedback = styled.div`
   display: flex;
