+++ packages/frontend/src/constants/routes.ts
@@ -1,0 +1,14 @@
+export enum InternalRoutes {
+  Home = '/',
+  About = '/about',
+  Feedback = '/feedback', // TODO: need to change
+  Legislator = '/lawmaker',
+  Topic = '/topics',
+  Speech = '/a',
+}
+
+export enum ExternalRoutes {
+  Medium = 'https://medium.com/twreporter',
+  Subscription = 'https://www.twreporter.org/account/email-subscription',
+  Support = 'https://support.twreporter.org/',
+}
